+++ mementoembed/version.py
@@ -1,3 +1,3 @@
 __appname__ = "MementoEmbed"
-__appversion__ = '0.2019.05.11.145955'
+__appversion__ = '0.2019.05.22.032641'
 __useragent__ = "{}/{}".format(__appname__, __appversion__)
